Lock more often, and add a test case for a common pattern
It is not uncommon to want one goroutine to turn some events `On`/`Off`
and to trigger those events, regardless of their current status,
from another goroutine.
This triggered the race detector, as synchronization for the map read/writes
was incomplete.
This adds a test which would fail before this changeset, and the fix to it.